@@ -7,7 +7,6 @@ using Microsoft.Extensions.Logging;
 using NBitcoin;
 using Swashbuckle.AspNetCore.Swagger;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using WalletWasabi.Backend.Middlewares;
@@ -63,7 +62,7 @@ namespace WalletWasabi.Backend
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
 #pragma warning disable IDE0060 // Remove unused parameter
 
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env, Global global)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env, Global global)
 #pragma warning restore IDE0060 // Remove unused parameter
 		{
 			app.UseStaticFiles();
@@ -81,7 +80,9 @@ namespace WalletWasabi.Backend
 
 			app.UseResponseCompression();
 
-			app.UseMvc();
+			app.UseRouting();
+
+			app.UseAuthorization();
 
 			var applicationLifetime = app.ApplicationServices.GetRequiredService<IApplicationLifetime>();
 			applicationLifetime.ApplicationStopping.Register(() => OnShutdown(global)); // Don't register async, that won't hold up the shutdown
